Initial refactoring and impl
