+++ src/main/scala/kubeyml/deployment/api/package.scala
@@ -78,7 +78,8 @@ package object api {
                 readinessProbe,
                 Map.empty
               )
-            )
+            ),
+            List.empty
           )
         ),
         strategy = RollingUpdate()
