+++ src/main/scala/kubeyml/ingress/json_support.scala
@@ -34,8 +34,6 @@ object json_support {
 
   implicit val apiVersionEncoder: Encoder[ApiVersion] = Encoder.instance(_.show.asJson)
 
-  implicit val hostEncoder: Encoder[Host] = Encoder.encodeString.contramap(_.value)
-
   implicit val pathEncoder: Encoder[Path] = Encoder.instance {
     case Path(ServiceMapping(serviceName, servicePort), path) =>
       Json.obj(
